fleet/requirements.txt
# Fleet Management Server Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client (for fleet agent)
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1

# System metrics (for fleet agent)
psutil>=5.9.0

